// erzahler/models/enumeration/turn-type-enum.ts
export enum TurnType {
  GAME_START = 'Game Start',
  ORDERS_AND_VOTES = 'Spring Orders/Votes',
  SPRING_ORDERS = 'Spring Orders',
  SPRING_RETREATS = 'Spring Retreats',
  SPRING_FINAL = 'Spring Final',
  FALL_ORDERS = 'Fall Orders',
  FALL_RETREATS = 'Fall Retreats',
  FALL_FINAL = 'Fall Final',
  ADJUSTMENTS = 'Winter Adjustments',
  ADJ_AND_NOM = 'Winter Adjustments/Nominations',
  ADJUSTMENTS_FINAL = 'Winter Final',
  NOMINATIONS = 'Nominations',
  VOTES = 'Votes'
}